third_party/WebKit/LayoutTests/virtual/mediasession/media/mediasession/mediaimage-expected.txt

CONSOLE WARNING: line 6: dummy log entry to avoid an empty -expected.txt in virtual/
This is a testharness.js-based test.
PASS Test that MediaImage is constructed using a dictionary 
PASS Test the different values allowed in MediaImage init dictionary 
PASS Test the default values for MediaImage 
PASS Test that passing unknown values to the dictionary is a no-op 
Harness: the test ran to completion.
